Make strncat depened on the size of dst
gcc was complaining about using the size of buf to delimit strncat,
which probably was fine in this case, since buf is gonna be smaller then
p, but fixing it hurts nobody, so do that.

Create a new Macro called EXPAND_SIZE, which just does x * CHAR_WIDTH,
and use that to determine the size of p (p is allocated with
EXPAND_SIZE(o->length) and o->length is later passed as size argument).