Don't assume ASCII on input strings
We used plain strlen(s) to determine the size of the input string (and
thus what we don't need to pad to reach the desired length), which does
not account for utf8 strings.

The expected of output of this command:
```
pad -m left -s ö -l 4 -c n
```
is nnnö, but it would have returned nö.

Replacing strlen() with utf8_strlen() from wee-utf8, the output becomes
nnö. It appears that we also truncated one character from the output
somehow, which can be fixed by adding one to the size argument of
padding().

That is possibly due to indexing, but I'm not sure.