Relicense my code to GPL-2.0-or-later and add str_buf
This adds str_buf, a struct based on the kernels seq_buf one. This also
means that I have to relicense my code from this point onwards as
GPL-2.0-or-later, so that I can actually use it. The Copyright in
common.h is attributed to the output of get_maintainers.pl
include/linux/seq_buf.h ; Hope that is correct?